Task-Url: http://github.com/shadedblue/campingbot/issues/101

+++ ca.hapke.campbinning.bot/src/ca/hapke/campingbot/callback/api/CallbackId.java
@@ -11,25 +11,25 @@ import ca.hapke.util.StringUtil;
 public class CallbackId {
 	private final String command;
 	private final int updateId;
-	private final int[] ids;
+	private final long[] ids;
 	private String result;
 
-	public CallbackId(String command, int updateId, List<Integer> ids) {
+	public CallbackId(String command, int updateId, List<Long> ids) {
 		this(command, updateId, unbox(ids));
 	}
 
-	private static int[] unbox(List<Integer> input) {
-		int[] output = new int[input.size()];
+	private static long[] unbox(List<Long> input) {
+		long[] output = new long[input.size()];
 
 		for (int j = 0; j < input.size(); j++) {
-			int k = input.get(j);
+			long k = input.get(j);
 			output[j] = k;
 
 		}
 		return output;
 	}
 
-	public CallbackId(String command, int updateId, int... ids) {
+	public CallbackId(String command, int updateId, long... ids) {
 		this.command = command;
 		this.updateId = updateId;
 		this.ids = ids;
@@ -43,10 +43,10 @@ public class CallbackId {
 
 		String command = words[0];
 		int messageId = Integer.parseInt(words[1]);
-		int[] ids = new int[words.length - 2];
+		long[] ids = new long[words.length - 2];
 		for (int i = 2; i < words.length; i++) {
 			String s = words[i];
-			ids[i - 2] = Integer.parseInt(s);
+			ids[i - 2] = Long.parseLong(s);
 		}
 		return new CallbackId(command, messageId, ids);
 	}
@@ -56,7 +56,7 @@ public class CallbackId {
 		inputs[0] = command;
 		inputs[1] = Integer.toString(updateId);
 		for (int i = 0; i < ids.length; i++) {
-			inputs[i + 2] = Integer.toString(ids[i]);
+			inputs[i + 2] = Long.toString(ids[i]);
 		}
 		return StringUtil.join(inputs, AbstractCommand.DELIMITER);
 	}
@@ -73,7 +73,7 @@ public class CallbackId {
 		return updateId;
 	}
 
-	public int[] getIds() {
+	public long[] getIds() {
 		return ids;
 	}
