+++ ca.hapke.campbinning.bot/src/ca/hapke/campingbot/voting/VoteCluster.java
@@ -11,9 +11,9 @@ import ca.hapke.campingbot.users.CampingUser;
  * @author Nathan Hapke
  */
 public class VoteCluster<T> {
-	private final Map<String, Map<CampingUser, Integer>> votesCluster = new HashMap<>();
+	private final Map<String, Map<CampingUser, Long>> votesCluster = new HashMap<>();
 	private final Map<String, Set<CampingUser>> votesNotApplicableCluster = new HashMap<>();
-	private final Map<String, Map<Integer, T>> valueMapCluster = new HashMap<>();
+	private final Map<String, Map<Long, T>> valueMapCluster = new HashMap<>();
 	private VoteTracker<T> tracker;
 
 	public VoteCluster(VoteTracker<T> voteTracker) {
@@ -26,7 +26,7 @@ public class VoteCluster<T> {
 		valueMapCluster.put(id, new HashMap<>());
 	}
 
-	public Map<CampingUser, Integer> getVotes() {
+	public Map<CampingUser, Long> getVotes() {
 		String key = tracker.createClusterKey();
 		return getVotes(key);
 	}
@@ -36,12 +36,12 @@ public class VoteCluster<T> {
 		return getVotesNotApplicable(key);
 	}
 
-	public Map<Integer, T> getValueMap() {
+	public Map<Long, T> getValueMap() {
 		String key = tracker.createClusterKey();
 		return getValueMap(key);
 	}
 
-	private Map<CampingUser, Integer> getVotes(String key) {
+	private Map<CampingUser, Long> getVotes(String key) {
 		if (!valueMapCluster.containsKey(key))
 			createCluster(key);
 		return votesCluster.get(key);
@@ -53,7 +53,7 @@ public class VoteCluster<T> {
 		return votesNotApplicableCluster.get(key);
 	}
 
-	private Map<Integer, T> getValueMap(String key) {
+	private Map<Long, T> getValueMap(String key) {
 		if (!valueMapCluster.containsKey(key))
 			createCluster(key);
 		return valueMapCluster.get(key);
